Update README.md
The Python Turtle Race

The Objective: The player whose turtle reaches its home first wins the game.

How to Play:

Each player rolls a dice to get a number.
The player then moves their turtle by that many steps.
The players alternate turns until one of them wins.
The Structure:

Each player had a turtle indicated by a different color. You can have more than two players, but for the sake of this tutorial, you’ll be creating a two-player game.
Each turtle has a home position that it must reach.
Each player uses a die to choose a value at random for their turn. In your program, the die is represented by a list of numbers from 1 to 6.

Developing the Game
It’s time to develop the code for the rest of the game. You’ll be using loops and conditional statements here, so you need to be careful with the indentations and spaces. To start, take a look at the steps your program will need to take to run the game:

Step 1: You’ll start by telling your program to check if either turtle has reached its home.
Step 2: If they haven’t, then you’ll tell your program to allow the players to continue trying.
Step 3: In each loop, you tell your program to roll the die by randomly picking a number from the list.
Step 4: You then tell it to move the respective turtle accordingly, with the number of steps based on the outcome of this random selection.

Conclusion
In this tutorial, you’ve learned how to program with the Python turtle library and grasped some very important programming concepts. You know how to deal with variable initialization, loops, conditional statements, indentations, lists, and operators. This is a great start for you, especially if you’re new to the Python programming language!